+++ backend/ticket-booking/src/main/java/com/booking/api/dto/TripCalendarPriceResponse.java
@@ -3,6 +3,7 @@ package com.booking.api.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -10,7 +11,7 @@ import java.time.LocalDate;
 public class TripCalendarPriceResponse {
 
     private LocalDate date;
-    private Double minPrice;
+    private BigDecimal minPrice;
     private boolean available;
 }
